refactor(analytics): move useCustomAnalyticsScript to lib/analytics/hooks.ts
Separates data-fetching from components: hooks.ts is the data layer,
shared.tsx is the component layer. Also converts CustomAnalyticsUpdateForm
to a default export per Next.js convention.
